explore -> access permission and enable location fix

+++ app/src/main/java/com/xborg/vendx/database/Location.kt
@@ -3,6 +3,6 @@ package com.xborg.vendx.database
 import com.squareup.moshi.Json
 
 data class Location(
-    @Json(name = "Longitude") var longitude: Float,
-    @Json(name = "Latitude") var latitude: Float
+    @Json(name = "Longitude") var longitude: Double,
+    @Json(name = "Latitude") var latitude: Double
     )
